Add files via upload
You are given a list of student names along with their scores in multiple subjects. Each student can have
scores in different subjects. Write a C++ program to store this information in a std::map where the key is the student name and the value is a map of subject names to
scores. Then, display the student names along with their total scores using the std::map.