Add files via upload
Inverted index of a book. Write a program that takes input and compiles an alphabetical index
of which words appear on which lines, as in the following input. Ignore case and punctuation.
It was the best of times,
it was the worst of times,
it was the age of wisdom,
it was the age of foolishness,
age 3-4
best 1
foolishness 4
it 1-4
of 1-4
the 1-4
times 1-2
was 1-4
wisdom 4
worst 2
Hint: create a map<string,vector<int>> whose key is a string that represents a word and whose value is
a vector<int> that represents the list of pages on which the word appears