Add files via upload
Write a program to convert a given string to title case. Use STL algorithm adjacent_find to find the
first two adjacent elements such that the first is space character and the second is a lower case alphabet.
Use std::toupper to convert the lower case alphabet to upper case.
Title case means that the first letter of each word is capitalized. For example, the string "hello world"
should be converted to "Hello World"